requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
groq==0.19.0
numpy
python-decouple
gunicorn
pypdf2
python-dotenv
whitenoise
dj-database-url
psycopg2-binary
requests
Pillow
djangorestframework-simplejwt